Add files via upload
Firefly Algorithm is applied to a 1D search space, where fireflies represent candidate solutions. The algorithm initialises a population of fireflies randomly within a specified domain. The movement of fireflies is determined by their attractiveness and is influenced by the intensity of their light (objective function value). Fireflies are attracted to brighter neighbours, and their movement is controlled by parameters such as attractiveness, step size, and dynamic alpha parameters.

The algorithm iteratively updates the positions of fireflies based on their attractiveness and the alpha parameter, aiming to converge towards an optimal solution. After a specified number of iterations, the algorithm identifies the firefly with the lowest objective function value, representing the optimal solution.

Code concludes by printing the optimal solution and its corresponding objective function value, and also provides a visualisation of the optimisation process, showing the objective function curve and marking the optimal solution